+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 ## [Unreleased]
 
+### Added
+- method, operatorPhone and operatorEmail parameters to PixInfraction resource
+
 ## [0.5.2] - 2024-10-18
 ### Fixed
 - IssuingCard expand expiration